Finish pcap-ng refactor, split remaining code into files
