+++ src/pcapng/header.rs
@@ -1,0 +1,12 @@
+#[derive(Debug)]
+pub struct PcapNGHeader {
+    pub magic_number: u32,
+    pub version_major: u16,
+    pub version_minor: u16,
+    pub thiszone: i32,
+    pub sigfigs: u32,
+    /// max len of captured packets, in octets
+    pub snaplen: u32,
+    /// Data link type
+    pub network: u32,
+}
