+++ src/pcapng/name_resolution.rs
@@ -5,9 +5,10 @@ use nom::multi::many_till;
 use nom::{Err, IResult};
 use rusticata_macros::{align32, newtype_enum};
 
+use crate::endianness::{PcapBE, PcapEndianness, PcapLE};
 use crate::{opt_parse_options, PcapError, PcapNGOption, NRB_MAGIC};
 
-use super::{ng_block_parser, PcapBE, PcapEndianness, PcapLE, PcapNGBlockParser};
+use super::*;
 
 #[derive(Clone, Copy, Eq, PartialEq)]
 pub struct NameRecordType(pub u16);
